backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==4.3.1
mysqlclient==2.2.0
channels==4.0.0
daphne==4.0.0
python-socketio==5.11.3
requests==2.31.0
PyYAML==6.0.1
flasgger==0.9.7.1
python-engineio==4.9.1
PySocks==1.7.1
beautifulsoup4==4.12.2
whitenoise==6.6.0
drf-spectacular==0.27.0
django-secure==1.0.2
python-socks==2.3.0
channels_redis==4.1.0
redis==5.0.1
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-prometheus==2.3.1
prometheus-client==0.19.0
sentry-sdk==1.39.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-cacheops==7.0.2
python-magic==0.4.27
django-cleanup==8.0.0
dnspython==2.4.2
dkimpy==1.1.5
email-validator==2.1.0.post1
django-channels-jwt-auth-middleware==1.0.0
django-csp==3.7
bleach==6.1.0
django-honeypot==1.2.1
django-debug-toolbar==4.2.0
django-silk==5.0.4
django-compression-middleware==0.5.0
django-querycount==0.8.3
django-storages==1.14.2
drf-yasg==1.21.7
drf-spectacular-sidecar==2024.2.1
markdown==3.5.2
  